+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Computation of the L2 norm of the pressure residual
 - Ability to solve the Poisson pressure equation via successive over-relaxation (SOR).
 - Computation of new U and V values based on pressure solution
+- "Inflow" grid preset, which is an empty fluid grid with upper and lower no-slip
+  boundaries, an inflow boundary on the left and an outflow boundary on the right.
 
 ## [0.1.0] - 2025-02-04